+++ src/packages/core/content-type/repository/content-type-structure-data-source.interface.ts
