+++ src/packages/core/content-type/repository/content-type-structure-repository.interface.ts
